Fix undelegating functionality
Currently, setting a KT1 address's delegate to 'undelegated' on TezBox fails. Undelegating is allowed by the Tezos protocol.

TezBox passes the empty string to `rpc.setDelegate` to indicate that the delegate should be cleared: https://github.com/tezbox/web-wallet/blob/master/app/controllers.js#L526

The protocol will only clear a delegate if it receives a delegation operation without a `delegate` field in the payload. Currently, eztz.js always sets a delegate. Instead, check if the empty string is passed and omit the field if this is true.